src/dwelling.pl: add library for handling dwelling places
Add a library called dwelling which handles where people have lived.
Define dwelt(Person, Place, Number, Source) to indicate that an
individual has dwelt at a particular place, or people_dwelt for a group
of people. This creates periods dwell(Person, Place, Number) which are
automatically contained within lifetime(Person).

Also included is moved(Person, From, FromNum, To, ToNum, Source) and
people_moved for a group of people. This creates a new event
[people_]move(Person/People, From, FromNum, To, ToNum), and also implies
[people_]dwell(Person/People, To, ToNum, Source). A move coincides with
the end of the From dwelling period and the beginning of the To dwelling
period.

Signed-off-by: James Hogan <james@albanarts.com>