src/chron.pl: add invalid event detection
When a constraint contains an event that isn't valid (doesn't exist)
apply_constraints_inner will fail resulting in repeated backtracking.
Add some failure handling to this code to produce a meaningful error
message when an invalid event is detected, and at the higher level when
the application of a constraint fails.

Signed-off-by: James Hogan <james@albanarts.com>