src/chron.pl: make apply_preargs/3 deterministic
Make apply_preargs/3 deterministic by refactoring it to use maplist so
that the differing atoms describing how to process the arguments (such
as event_time/1 and pass/1) are at the top level of the arguments,
allowing them to be indexed and to avoid choices. Also remove the cut
while we're at it.

Reported-by: Markus Triska <markus.triska@gmx.at>
Signed-off-by: James Hogan <james@albanarts.com>