src/precompute.pl: add code to precompute a predicate
Add code to precompute a predicate according to a mapping rule. This is
useful for both efficiency, determinism, and removing duplicates.

Use precompute_mapping/2 to define a precompute mapping to be computed
by precompute/0. precompute/2 can also be used to explicitly precompute
a particular mapping.

A simple example would be:
:- precompute(is_person(Person), person(Person)).

Which creates is_person/1 to match person/1.

Signed-off-by: James Hogan <james@albanarts.com>