src/test.pl: add test harness
Add a simple test library for running tests, gathering results, and
finally printing results. It isn't particularly complicated or powerful
at the moment, but it does the trick.

Each test has a set of flags, which are actually processed using
result_check/3. This takes the flag itself, a list of the flag for each
solution (allowing specific unifications to be used by result_check/3),
and an output result which is stored and used to print a status at the
end.

Some flags are defined:
 - results(N): specifies the number of solutions to be expected
 - single: specifies that a single solution is expected
 - msg(Pass, Fail): simply display a pass or fail message depending on
   whether a single solution was found
 - cond(Condition, Pass, Fail): test a condition and print a pass or
   fail message
 - shell(Command, Pass, Fail): run a shell command and print a pass or
   fail message based on the exit status
 - file_match(File1, File2): diffs two files with diff -q and fails if
   they don't match identically

Signed-off-by: James Hogan <james@albanarts.com>