src/*.pl: convert prolog code to use modules
Convert most prolog files to modules which export various predicates.

We rely on "-p chron=src/" being passed on the swipl command line so
that chron modules can be referred to as chron(Module) when including or
use_module'ing them.

Modules which want to become part of the chron database call
chron:database(Module). This is done automatically (as well as importing
chron and clpfd) by including the new wrapper script chron(database).

Add an import_ns_predicates/1 to the chron module which effectively
imports certain predicates from database module namespaces into the
corresponding predicates in the chron module. This is used to make chron
data in different modules visible to the chron code. Since this defines
another instance of each procedure, the no-op rules (":- fail") can be
removed.

Add wrappers for several of these imported predicates that can be
exported from chron, so that predicates for defining facts and
predicates for testing the database don't conflict. These are created
automatically with an is_ prefix by the wrap_ns_predicates/1 call. These
may be converted to precomputed predicates later.

Signed-off-by: James Hogan <james@albanarts.com>