Makefile.build: fix link order with --as-needed
When pkg-config adds --as-needed to the library link command line we get
lots of undefined symbols. This is because it appears to require
referenced libraries to appear later in the linker command line than the
referencers.

Move the local object files before the libraries on the linker command
line to resolve this problem.

Link: https://wiki.gentoo.org/wiki/Project:Quality_Assurance/As-needed#Importance_of_linking_order
Fixes 0bdd423c99fb ("Add a Makefile build system")
Signed-off-by: James Hogan <james@albanarts.com>