Makefile.build: separate ldflags from cflags
When using pkg-config to get the compiler flags (--cflags)  and linker
flags (--libs) for a particular library, get them separately and pass
only the compiler flags when compiling object files, and only the linker
flags when linking the final binary.

There is no need to bloat the command lines, and it was making the make
output noisy too.

Fixes 0bdd423c99fb ("Add a Makefile build system")
Signed-off-by: James Hogan <james@albanarts.com>